boxmap_gpu works

+++ src/algorithms.jl
@@ -1,6 +1,6 @@
-function relative_attractor(F::BoxMap, B::BoxSet{<:AbstractBoxPartition{Box{N,T}}}; steps=12) where {N,T}
-    for k = 1:steps
-        B = subdivide(B, (k%N)+1)
+function relative_attractor(F::BoxMap, B::BoxSet{<:AbstractBoxPartition{Box{N,T}}}; steps=12i32) where {N,T}
+    for k = 1i32:steps
+        B = subdivide(B, (k % N) + 1i32)
         B = B ∩ F(B)
     end
     return B
@@ -16,9 +16,9 @@ function unstable_set!(F::BoxMap, B::BoxSet)
     return B
 end
 
-function chain_recurrent_set(F::BoxMap, B::BoxSet{<:AbstractBoxPartition{Box{N,T}}}; steps=12) where {N,T}
-    for k in 1:steps
-        B = subdivide(B, (k%N)+1)
+function chain_recurrent_set(F::BoxMap, B::BoxSet{<:AbstractBoxPartition{Box{N,T}}}; steps=12i32) where {N,T}
+    for k in 1i32:steps
+        B = subdivide(B, (k % N) + 1i32)
         P = TransferOperator(F, B)
         B = strongly_connected_components(P)
     end
@@ -36,10 +36,10 @@ end
     end
     h = armijo_rule(g, x, 1f0, 1f-4, 0.8f0)
 
-    expon(ϵ, σ, h, δ) = Int(ceil(log(ϵ * (1/2)^σ)/log(maximum((1 - h, δ)))))
+    expon(ϵ, σ, h, δ) = Int(ceil(log(ϵ * (1/2)^σ)/log(maximum((1i32 - h, δ)))))
     n = expon(0.2f0, k, h, 0.1f0)
 
-    for _ in 1:n
+    for _ in 1i32:n
         Dg = g_jacobian(x)
         x = x - h * (Dg \ g(x))
     end
@@ -47,10 +47,10 @@ end
     return x
 end
 
-function cover_roots(g, Dg, B::BoxSet{<:AbstractBoxPartition{Box{N,T}}}; steps::Int=12) where {N,T}
+function cover_roots(g, Dg, B::BoxSet{<:AbstractBoxPartition{Box{N,T}}}; steps=12i32) where {N,T}
     domain = B.partition.domain
-    for k in 1:steps
-        B = subdivide(B, (k%N)+1)
+    for k in 1i32:steps
+        B = subdivide(B, (k % N) + 1i32)
         f = x -> adaptive_newton_step(g, Dg, x, k)
         F_k = BoxMap(f, domain, no_of_points = 40)
         B = F_k(B)
@@ -113,7 +113,7 @@ const half, sixth, third = Float32.((1/2, 1/6, 1/3))
     return @. x + τ * dx
 end
 
-@propagate_inbounds function rk4_flow_map(v, x; step_size=0.01f0, steps=Int32(20))
+@propagate_inbounds function rk4_flow_map(v, x; step_size=0.01f0, steps=20i32)
     for _ in 1:steps
         x = rk4(v, x, step_size)
     end
